Working on fixing up use of masks throughout, but issues remain.
The mask8/mask16 vs. m256/m512 representing bool masks gives us the same
problems that come up with std::vector<bool>, where for example mask8
is just 8 bits, not a vector of 8 'ints'. This breaks everything which
uses the algorithms or array indexing/iteration on the vbool* on
Skylakes. In the end we probably will need to do something like
std::vector<bool> for these masks, and abstract some from the SSE/AVX
style of mask, which uses a whole vector register and the AVX512 style
of mask which uses a char or short bitmask.